Resolves two conflicts in src/AiModelBuilder.BuildPipeline.cs:

1. ShouldUseDirectSupervisedNeuralTraining (semantic contradiction): the PR routed ALL
   transformers to direct neural training unconditionally, on the premise that the optimizer
   path leaves a plain transformer at chance. master's newer commit refuted that premise as a
   MEASUREMENT artifact (streamed MSE-fitness + double-softmaxed perplexity, not real learning)
   and VERIFIED the optimizer path trains transformers correctly (chance -> ~100% / ~89% on
   held-out top-1). Resolved to master's gated version (route to direct training only when
   mixed-precision / memory config is set) so the merge does not regress master's verified fix
   or forfeit the optimizer's batching/scheduler/early-stop for plain transformers.

2. Epoch-hook optimizer branch: both branches carry master's #1790 per-epoch callback bridge;
   the PR additionally wraps the epoch-hook finalOptimizer.Optimize call in a zero-alloc
   TensorArena scope (#1804) — this PR's actual purpose. Kept the PR's arena version (= master's
   bridge + the arena line). All 3 TensorArena.Create() training scopes preserved.

TimeSeriesModelBase.cs auto-merged. Builds clean on net10.0 + net471 (src + tests).

+++ tests/AiDotNet.Tests/UnitTests/Diffusion/Models/FastGenContractTests.cs
@@ -567,7 +567,7 @@ public class FastGenContractTests : DiffusionUnitTestBase
 
         Assert.NotNull(clone);
         Assert.NotSame(model, clone);
-        Assert.Equal(model.ParameterCount, (int)clone.ParameterCount);
+        Assert.Equal(model.ParameterCount, clone.ParameterCount); // long==long: (int) truncated valid >2.1B (e.g. 12B FLUX) counts
     }
 
     // FP32 (production-canonical for FLUX): a materialized ~12B-param FluxSchnell at FP64 would dwarf
@@ -594,7 +594,7 @@ public class FastGenContractTests : DiffusionUnitTestBase
 
         Assert.NotNull(clone);
         Assert.NotSame(model, clone);
-        Assert.Equal(model.ParameterCount, (int)clone.ParameterCount);
+        Assert.Equal(model.ParameterCount, clone.ParameterCount); // long==long: (int) truncated valid >2.1B (e.g. 12B FLUX) counts
     }
 
     #endregion
